Update style.css
I have updated the background colour in the css file using a multi coloured gradient feature. It should get reflected in the pages. Please check and let me know if the colour needs to be changed or any other changes need to be done. Also, if you merge, please accept it under the label hacktoberfest2023